server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
elevenlabs==2.7.1
fastapi==0.115.14
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.4
fsspec==2025.5.1
gotrue==2.12.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.7
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.7.0
moviepy==2.2.1
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai==1.93.0
openai-whisper==20250625
opencv-python-headless==4.10.0.84
packaging==25.0
pillow==11.2.1
pillow_heif==0.22.0
pluggy==1.6.0
postgrest==1.1.1
proglog==0.1.12
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.5.3
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
roboflow==1.1.66
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
websockets==15.0.1
stripe==10.11.0
shapely==2.0.6
boto3==1.40.15
pydicom==2.4.4
reportlab==4.1.0
playwright==1.49.0
xhtml2pdf==0.2.17
